+++ src/utils/pdf/generate/layouts/default.ts
@@ -1,8 +1,9 @@
 import jsPDF from 'jspdf'
 import { getVisibleReminders } from 'utils/reminderUtils'
-import PdfLayout, { reorderReminders } from 'utils/pdf/generate/layouts/layoutUtils'
-import { Logo } from 'utils/pdf/generate/logo'
+import { reorderReminders } from 'utils/reorder'
 import { titleCase } from 'utils/textUtils'
+import PdfLayout from 'utils/pdf/generate/layouts/layoutUtils'
+import { Logo } from 'utils/pdf/generate/logo'
 import { TPdfStyles, IPrintPdf } from 'types/pdf'
 
 const Styles: TPdfStyles = {
